+++ src/init.c
@@ -10,15 +10,17 @@
 
 /* .Call calls */
 extern SEXP ages_to_interval(SEXP, SEXP);
+extern SEXP breaks_to_interval(SEXP);
 extern SEXP cut_ages(SEXP, SEXP);
 extern SEXP aggregate_age_counts(SEXP, SEXP, SEXP);
 extern SEXP split_interval_counts(SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP reaggregate_interval_counts(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 
 static const R_CallMethodDef CallEntries[] = {
-    {"ages_to_interval",  (DL_FUNC) &ages_to_interval,  2},
-    {"cut_ages",  (DL_FUNC) &cut_ages,  2},
-    {"aggregate_age_counts",  (DL_FUNC) &aggregate_age_counts,  3},
+    {"ages_to_interval",  (DL_FUNC) &ages_to_interval, 2},
+    {"breaks_to_interval",  (DL_FUNC) &breaks_to_interval, 1},
+    {"cut_ages",  (DL_FUNC) &cut_ages, 2},
+    {"aggregate_age_counts",  (DL_FUNC) &aggregate_age_counts, 3},
     {"split_interval_counts", (DL_FUNC) &split_interval_counts, 5},
     {"reaggregate_interval_counts", (DL_FUNC) &reaggregate_interval_counts, 6},
     {NULL, NULL, 0}